breastw {dprep}	R Documentation

The Breast Wisconsin dataset

Description

This is the Breast Wisconsin dataset from the UCI Machine Learning Repository. Sixteen instances with missing values have been deleted from the original dataset.

Usage

data(breastw)

Format

A data frame with 683 observations on the following 10 variables.

V1

Clump Thickness

V2

Uniformity of Cell Size

V3

Uniformity of Cell Shape

V4

Marginal Adhesion

V5

Single Epithelial Cell Size

V6

Bare Nuclei

V7

Bland Chromatin

V8

Normal Nucleoli

V9

Mitoses

V10

Class: 1 for benign, 2 for Malign

Details

All the features assume values in the range 1-10. The original dataset contains 699 observations but 16 of them have been delate because contain missing values

Source

The UCI Machine Learning Database Repository at:

ftp://ftp.ics.uci.edu/pub/machine-learning-databases
http://www.ics.uci.edu/~mlearn/MLRepository.html

Examples

#Detecting outliers in class-1 using the LOF algorithms---
data(breastw)
maxlof(breastw[breastw[,10]==1,],name="",30,40)

[Package dprep version 2.1 Index]