ce.knn.imp {dprep}	R Documentation

Function that calls ec.knnimp to perform knn imputation

Description

This function simply sets up the dataset so that missing values can be imputed by knn imputation. ec.knnimp is the function that actually carries out the imputation.

Usage

ce.knn.imp(m, natr = rep(0, 0), k1)

Arguments

m 	 matrix containing relevant variables and classes

natr 	 list of nominal attributes

k1 	 number of neighbors to use for imputation

Value

r 	 matrix with missing values imputed

Author(s)

Caroline Rodriguez and Edgar Acuna

Examples

data(hepatitis)
hepa.knnimp=ce.knn.imp(hepatitis,natr=c(1,3:14),k1=10)

[Package dprep version 2.1 Index]