chiMerge {dprep}	R Documentation

Discretization using the Chi-Merge method

Description

This function performs supervised discretization using the Chi Merge method.

Usage

chiMerge(data, varcon, alpha = 0.1)

Arguments

data 	 The name of the dataset to be discretized

varcon 	 Vector of continuous variables

alpha 	 The significance level

Details

In case of datasets containing negative values apply first a range normalization to change the range of the attributes values to an interval containing positive values. The discretization process becomes slow when the number of variables increases (say for more than 100 variables).

Value

discdata 	 A new data matrix containing the discretized features

Author(s)

Edgar Acuna, Jaime Porras, and Carlos Lopez

References

Kantardzic M. (2003). Data Mining: Concepts, Models, methods, and Algorithms. John Wiley. New York.

See Also

disc.ef, disc.ew,disc.1r,disc.mentr

Examples

#-----Discretization using the ChiMerge method
data(my.iris)
iris.disc=chiMerge(my.iris,1:4,alpha=0.05)
#-----Applying chiMerge a dataset containing negative values 
#data(ionosphere)
#normionos=rangenorm(ionosphere,"mmnorm")
#ionos.disc=chiMerge(normionos,1:32)

[Package dprep version 2.1 Index]