circledraw {dprep}	R Documentation

circledraw

Description

This function draws a circle using the polygon function from the graphics package. It is an auxiliary function used by radviz2d.

Usage

circledraw (numpts = 200, radius = 1)

Arguments

numpts 	 Number of edges of the polygon, default is 200.

radius 	 Radius of the circle to be drawn, default is 1.

Details

A circle of a specified radius is drawn by the polygon function of the graphics library by constructing a polygon with numpts number of edges. It is intended to be an auxiliary function for the radviz2d visualization.

Value

Displays a circle of radius = radius.

Author(s)

Caroline Rodriguez

Examples

#----Circledraw examples
circledraw()

[Package dprep version 2.1 Index]