closest {dprep}	R Documentation

Auxiliary function used in the function baysout

Description

Function used by baysout to select the k vectors that are closer to a given instance.

Usage

closest(dis, neigh, k)

Arguments

dis 	 An instance from the dataset under study

neigh 	 A matrix containing the distance from the given observations to each of its k neighbors.

k 	 The number of nearest neighbors

Author(s)

Caroline Rodriguez

See Also

baysout

[Package dprep version 2.1 Index]