colon {dprep}	R Documentation

Alon et al.'s colon dataset

Description

This is Alon et al.'s Colon cancer dataset which contains information on 62 samples for 2000 genes. The samples belong to tumor and normal colon tissues.

Usage

data(colon)

Format

A data frame with 62 observations for 2000 genes. An additional column contains the tissue classes.

Source

The data is available at:

http://microarray.princeton.edu/oncology/

References

Alon U, Barkai N, Notterman DA, Gish, K, Ybarra, S. Mack, D and Levine, AJ. (1999). Broad patterns of gene expression revealed by clustering analysis of tumor and normal colon tissues probed by oligonucleotide arrays. Proc. Natl. Acad. Sci. USA. 96. p. 6745-6750.

Examples

#Detecting the top 5 outliers in class-2 using the LOF algorithm
data(colon)
colon2.lof=maxlof(colon[colon[,2001]==2,],"colon-class2")
colon2.lof[order(colon2.lof,decreasing=TRUE)][1:5]

[Package dprep version 2.1 Index]