combinations {dprep}	R Documentation

Constructing distinct permutations

Description

A function for constructing the minimal set of permutations of the elements in the sequence 1:numcol as described by Wegman in Hyperdimensional Data Analysis(1999)

Usage

combinations(numcol)

Arguments

numcol 	 A value representing the number of columns in a matrix

Value

A matrix in which each column represents a distinct permutation of the sequence 1:numcol

Author(s)

Caroline Rodriguez

References

Wegman, E. (1990), Hyperdimensional data analysis using parallel coordinates, Journal of the American Statistical Association, 85, 664-675.

[Package dprep version 2.1 Index]