cv10log {dprep}	R Documentation

10-fold cross validation estimation error for the classifier based on logistic regression

Description

10-fold cross validation estimation of the misclassification error for the classifier based on logistic regression

Usage

cv10log(data, repet,maxwts=2500)

Arguments

data 	 The name of the dataset

repet 	 The number of repetitions

maxwts 	 The maximum number of weights to be estimated. It must be an integer greater than the number of predictors of the dataset.

Value

The mean cross validation error for the classifier based on logistic regression using the number of repetitions

Author(s)

Edgar Acuna

References

Ripley, B.D. (1996). Pattern recognition and Neural networks. Cambridge University Press

Venables,W.N., and Ripley, B.D. (2002). Modern Applied Statistics with S. Fourth edition, Springer

See Also

crossval, cv10mlp

Examples

#-----cross validation error for the logistic classifier-------
data(bupa)
cv10log(bupa,5)

[Package dprep version 2.1 Index]