cv10mlp {dprep}	R Documentation

10-fold cross validation error estimation for the multilayer perceptron classifier

Description

10-fold cross validation estimation error for the multilayer perceptron classifier.

Usage

cv10mlp(data, units, decay = 0, maxwts = 1000, maxit = 100, 
repet)

Arguments

data 	 The name of the dataset

units 	 The number of units in the hidden layer

decay 	 The decay parameter

maxwts 	 The maximum number of weights to be estimated in the network

maxit 	 The maximum number of iterations

repet 	 The number of repetitions

Value

Returns the mean cross validation for the multilayer perceptron classifier.

Author(s)

Edgar Acuna

References

Ripley, B.D. (1996). Pattern recognition and Neural networks. Cambridge University Press

Venables,W.N., and Ripley, B.D. (2002). Modern Applied Statistics with S. Fourth edition, Springer

See Also

crossval, cv10log

Examples

#-----cross validation using the MLP classifier---
data(heartc)
cv10mlp(heartc,25,decay=0.1,maxwts=1000,maxit=100,repet=2)

[Package dprep version 2.1 Index]