cv10rpart2 {dprep}	R Documentation

Auxiliary function for sequential feature selection

Description

This function finds the number of instances correctly classified by the decision tree classifier, rpart, using 10-fold cross validation and one repetition.

Usage

cv10rpart2(datos)

Arguments

datos 	 The name of the dataset

Author(s)

Edgar Acuna

See Also

crossval

[Package dprep version 2.1 Index]