decscale {dprep}	R Documentation

Decimal Scaling

Description

This is a function to apply decimal scaling to a matrix or dataframe. Decimal scaling transforms the data into [-1,1] by finding k such that the absolute value of the maximum value of each attribute divided by 10^k is less than or equal to 1.

Usage

decscale(data)

Arguments

data 	 The dataset to be scaled

Details

Uses the scale function found in the R base package.

Value

decdata 	 The original matrix that has been scaled by decimal scaling

Author(s)

Caroline Rodriguez and Edgar Acuna

Examples

data(sonar)
def=par(mfrow=c(2,1))
plot(sonar[,2])
dssonar=decscale(sonar)
plot(dssonar[,2])
par(def)

[Package dprep version 2.1 Index]