disc.ew {dprep}	R Documentation

Discretization using the equal width method

Description

Unsupervised discretization using intervals of equal width. The widths are computed using Scott's formula.

Usage

disc.ew(data, varcon)

Arguments

data 	 The name of the dataset containing the attributes to be discretized

varcon 	 A vector containing the indexes of the attributes to be discretized

Value

Returns a new data matrix with discretized values.

Author(s)

Edgar Acuna

References

Venables, W.N., and Ripley, B.D. (2002). Modern Applied Statistics with S. Fourth edition, Springer

See Also

disc.ef, disc.1r,chiMerge,disc.mentr

Examples

#----Discretization using the equal frequency method
data(bupa)
bupa.disc.ew=disc.ew(bupa,1:6)

[Package dprep version 2.1 Index]