disc.mentr {dprep}	R Documentation

Discretization using the minimum entropy criterion

Description

This function discretizes the continuous attributes of a data frame using the minumum entropy criterion with the minimum description length as stopping rule.

Usage

disc.mentr(data, vars)

Arguments

data 	 The name of the dataset to be discretized

vars 	 A vector containing the indices of the columms to be discretized and column containing the classes

Value

Returns a matrix containing only discretized features.

Author(s)

Luis Daza

References

Dougherty, J., Kohavi, R., and Sahami, M. (1995). Supervised and unsupervised discretization of continuous features. ML-95.

See Also

disc.1r, disc.ew,disc.ef,chiMerge

Examples

data(my.iris)
iris.discme=disc.mentr(my.iris,1:5)

[Package dprep version 2.1 Index]