discretevar {dprep}	R Documentation

Performs Minimum Entropy discretization for a given attribute

Description

This function carries out ME discretization for a given attribute of a dataset. It is also called from within the function discr.mentr.

Usage

discretevar(data, var, n, p)

Arguments

data 	 The name of the dataset

var 	 The column where the attribute to be discretized is located

n 	 The number of rows of the dataset

p 	 The number of columns of the dataset

Author(s)

Luis Daza

See Also

disc.mentr

[Package dprep version 2.1 Index]