dist.to.knn {dprep}	R Documentation

Auxiliary function for the LOF algorithm.

Description

This function returns an object in which columns contain the indices of the first k neighbors followed by the distances to each of these neighbors.

Usage

dist.to.knn(dataset, neighbors)

Arguments

dataset 	 The name of the dataset

neighbors 	 The number of neighbors

Author(s)

Caroline Rodriguez

See Also

maxlof

[Package dprep version 2.1 Index]