dprep-package {dprep}	R Documentation

Data Preprocessing for Supervised Classification

Description

Functions for normalization, treatment of missing values, discretization, outlier detection, feature selection, and visualization

Details

Package: 	 dprep

Type: 	 Package

Version: 	 2.0

Date: 	 2008-04-14

License: 	 GPL (version 2 or later) See file LICENCE.

For an overview of how to use the package, see user manual provided.

Author(s)

Maintainer: Edgar Acuna <edgar@cs.uprm.edu>

[Package dprep version 2.1 Index]