eje1dis {dprep}	R Documentation

Basic example for discriminant analysis

Description

This data frame contains information about 32 students. The first two columns contain their grades obtained on the first two exams and the last column of the dataset contains the classes: P=Pass, and F=Fail

Usage

data(eje1dis)

Format

A data frame with 32 observations on the following 3 variables.

V1

Grade on the first exam

V2

Grade on the second exam

class

The class vector:P=Pass, F=Fail

Source

Data obtained from Edgar Acuna:

http://math.uprm.edu/~edgar/datosclass.html

Examples

#---- Performing 10-fold cross validation using the LDA classifier-----
data(eje1dis)
crossval(eje1dis,10,"lda",repet=5)

[Package dprep version 2.1 Index]