finco {dprep}	R Documentation

FINCO Feature Selection Algorithm

Description

This function selects features using the FINCO algorithm. The dataset must contain only discretized values.

Usage

finco(data,level)

Arguments

data 	 Name of the dataset containing the discretized values

level 	 Minimum inconsistency level

Details

The level value must be greater than the inconsistency of the whole dataset, which first must be discretized. The function inconsist included in this library computes inconsistencies. A small value for level yields a greater number of selected features.

Value

varselec 	 Index of selected features

inconsis 	 Inconsistency rates of the selected features

Author(s)

Edgar Acuna

References

Acuna, E , (2003) A comparison of filters and wrappers for feature selection in supervised classification. Proceedings of the Interface 2003 Computing Science and Statistics. Vol 34.

Acuna, E., Coaquira, F. and Gonzalez, M. (2003). A comparison of feature selection procedures for classifiers based on kernel density estimation. Proc. of the Int. Conf. on Computer, Communication and Control technologies, CCCT03. VolI. p. 468-472. Orlando, Florida.

See Also

inconsist,lvf

Examples

#---- Feature Selection with FINCO
data(my.iris)
iris.discew=disc.ew(my.iris,1:6)
inconsist(iris.discew)
finco(iris.discew,0.05)

[Package dprep version 2.1 Index]