hawkins {dprep}	R Documentation

The Hawkins-Bradu-Kass dataset

Description

An artificial dataset generated by Hawkins, Bradu, and Kass used for illustrating some of the merits of robust techniques.

Usage

data(hawkins)

Format

A data frame consisting of 75 observations on the following 4 variables.

V1

First predictor variable

V2

Second predictor variable

V3

Third predictor variable

V4

The response variable

Source

The data appears on p. 94 of Rousseeuw, P, and Leroy, A. (1987). Robust Regression and outlier detection. John Wiley & Sons. New York.

References

Hawkins, D.M, Bradu, D., Kass, G.V.(1984). Location of several outliers in multiple regression data using elemental sets. Technometrics, 26. 197-208.

Examples

#---- Finding outliers using the LOF algorithm----
data(hawkins)
haw.lof=maxlof(hawkins[,1:3],"Hawkins")
haw.lof[order(haw.lof,decreasing=TRUE)]

[Package dprep version 2.1 Index]