heartc {dprep}	R Documentation

The Heart Cleveland dataset

Description

This dataset contains information concerning heart disease diagnosis. The data was collected from the Cleveland Clinic Foundation, and it is available at the UCI machine learning Repository. Six instances containing missing values have been deleted from the original dataset.

Usage

data(heartc)

Format

A data frame with 297 observations on the following 14 variables.

V1

age(continuous)

V2

sex

V3

cp, chest pain type:1,2,3,4

V4

trestbps: resting blood pressure(continuous)

V5

cholesterol(continuous)

V6

fps: fatsing blood sugar>120? yes=1, no =0

V7

restecg: resting electrocardiographic results, 0,1, 2

V8

thalach: maximum heart rate achieved(continuous)

V9

exang: exercise induced angina (1 = yes; 0 = no)

V10

oldpeak = ST depression induced by exercise relative to rest (continuous)

V11

slope: the slope of the peak exercise ST segment

V12

ca: number of major vessels (0-3) colored by flourosopy

V13

thal: 3 = normal; 6 = fixed defect; 7 = reversable defect

V14

diagnosis of heart disease: 1: < 50 2: > 50

Details

Six instances containing missing values have been deleted from the original dataset. This dataset includes continuous, binomial, nominal, and ordinal features.

Source

The UCI Machine Learning Database Repository at:

ftp://ftp.ics.uci.edu/pub/machine-learning-databases
http://www.ics.uci.edu/~mlearn/MLRepository.html

Examples

#----Detecting outliers using the Relief---
data(heartc)
relief(heartc,100,0.4,vnom=c(2,3,6,7,9,11:13))

[Package dprep version 2.1 Index]