hepatitis {dprep}	R Documentation

The hepatitis dataset

Description

This is the hepatitis dataset from the UCI. The data was donated by Gail Gong.

Usage

data(hepatitis)

Format

A data frame with 155 observations on the following 20 variables. This dataset contains a large number of missing values.

V1

Histology:no,yes

V2

age

V3

sex: male,female

V4

steroid:no,yes

V5

antivirals:no,yes

V6

fatigue:no, yes

V7

malaise:no, yes

V8

anorexia:no, yes

V9

liver big:no,yes

V10

liver firm:no,yes

V11

spleen palpable: no, yes

V12

spiders:no,yes

V13

ascites:no,yes

V14

Varices:no,yes

V15

Bilirubin

V16

alk phosphate

V17

sgot

V18

Albumin

V19

Protime

V20

Class:Die, Live

Details

The original dataset has the class labels in the first column.

Source

The UCI Machine Learning Database Repository at:

ftp://ftp.ics.uci.edu/pub/machine-learning-databases
http://www.ics.uci.edu/~mlearn/MLRepository.html

References

Diaconis,P. & Efron,B. (1983). Computer-Intensive Methods in Statistics. Scientific American, Volume 248.

Examples

#------Report and plot of missing values ------
data(hepatitis)
imagmiss(hepatitis,"Hepatitis")

[Package dprep version 2.1 Index]