inconsist {dprep}	R Documentation

Computing the inconsistency measure

Description

This function computes the inconsistency of a discretized dataset.

Usage

inconsist(data)

Arguments

data 	 a discretized dataset

Details

This function requires the function row.matches included in this environment package, and the function unique from the base library.

Value

incon 	 the inconsistency measure of the dataset

Author(s)

Edgar Acuna

References

Dash M., Liu H, and Motoda, H. (1998). Consistency Based Feature Selection Pacific-Asia Conference on Knowledge Discovery and Data Mining

See Also

finco, lvf

Examples

##---- Calculating Inconsistency ----
data(bupa)
bupa.discew=disc.ew(bupa,1:6)
inconsist(bupa.discew)

[Package dprep version 2.1 Index]