ionosphere {dprep}	R Documentation

The Ionosphere dataset

Description

The Ionosphere dataset from the UCI Machine Learning Repository

Usage

data(ionosphere)

Format

A data frame with 351 observations on the following 33 variables.

Details

The original dataset contains 34 predictors, but we have eliminated the two first features, because the first feature had the same value in one of the classes and the second feature assumes the value 0 in all observations.

Source

The UCI Machine Learning Database Repository at:

ftp://ftp.ics.uci.edu/pub/machine-learning-databases
http://www.ics.uci.edu/~mlearn/MLRepository.html

Examples

#---Outlier detection in  ionosphere class-1 using the Mahalanobis distiance----
data(ionosphere)
mahaout(ionosphere,1)

[Package dprep version 2.1 Index]