lvf {dprep}	R Documentation

Las Vegas Filter

Description

Las Vegas Filter uses a random generation of subsets and an inconsistency measure as the evaluation function to determine the relevance of features in the dataset.

Usage

lvf(data, lambda, maxiter)

Arguments

data 	 Name of the discretized dataset

lambda 	 Threshold for the inconsistency

maxiter 	 Maximum number of iterations

Details

If the dataset has continuous variables, these must first be discretized. This package includes four discretization methods. A value of lambda close to the inconsistency of the whole dataset yields a large number of selected features, a large lambda yields few selected features.

Value

bestsubset 	 The best subset of features

Author(s)

Edgar Acuna

References

LIU, H. and SETIONO, R. (1996). A probabilistic approach to feature selection: a filter solution. Proc. of the thirteenth International Conference of Machine Learning, 319-337.

See Also

disc.ew,inconsist,finco

Examples

#---- LVF method ----
data(my.iris)
iris.discew=disc.ew(my.iris,1:6)
inconsist(iris.discew)
lvf(iris.discew,0,300)

[Package dprep version 2.1 Index]