mahaout {dprep}	R Documentation

Multivariate outlier detection through the boxplot of the Mahalanobis distance

Description

This function finds multivariate outliers by constructing a boxplot of the Mahalanobis distance of all the instances.

Usage

mahaout(data, nclass, plot = TRUE)

Arguments

data 	 Name of the dataset

nclass 	 Number of the class to check for outliers

plot 	 Logical value. If plot=T a plot of the mahalanobis distance is drawn

Details

uses cov.rob function from the MASS library

Value

Returns a list of top outliers according to their Mahalanobis distance and a list of all the instances ordered according to their Mahalanobis distance.
If Plot=T, a plot of the instances ranked by their Mahalanobis distance is provided.

Author(s)

Edgar Acuna

References

Rousseeuw, P, and Leroy, A. (1987). Robust Regression and outlier detection. John Wiley & Sons. New York.

See Also

robout

Examples

#---- Detecting outliers using the Mahalanobis distance----
data(bupa)
mahaout(bupa,1)

[Package dprep version 2.1 Index]