mardia {dprep}	R Documentation

The Mardia's test of normality

Description

Performs the Mardia's test to check for multivariate normality

Usage

mardia(data)

Arguments

data 	 The dataset containing the features for which multivariate normality is going to be tested. The last column contains the class. In case of unsupervised data add a dummy colummn of ones. In case of regression data, transform the response column in a column of ones

Value

Returns the p-values for the corresponding third and fourth moments of the multivariate normal distribution.

Author(s)

Edgar Acuna

References

Mardia, K.V. (1985). "Mardia's Test of Multinormality," in S. Kotz and N.L. Johnson, eds., Encyclopedia of Statistical Sciences, vol. 5 (NY: Wiley), pp. 217-221.

See Also

vvalen

Examples

#-----Mardia test for supervised data-----
data(my.iris)
mardia(my.iris)
#----Mardia test for unsupervised data-----
data(hawkins)
haw=cbind(hawkins[,-4],rep(1,75))
mardia(haw)

[Package dprep version 2.1 Index]