maxdist {dprep}	R Documentation

Auxiliary function used when executing the Bay's algorithm for outlier detection

Description

This function is used by the function baysout in this package, to find the largest value of a distance vector. Returns the value and the index number of the largest distance.

Usage

maxdist(dneighbors)

Arguments

dneighbors 	 The value and the index number of the largest distance

Author(s)

Caroline Rodriguez

See Also

baysout

[Package dprep version 2.1 Index]