mo4 {dprep}	R Documentation

The fourth moment of a multivariate distribution

Description

This function computes the fourth moment of a multivariate distribution. This result is used later in the mardia's test for multivariate normality.

Usage

mo4(data)

Arguments

data 	 The dataset containing the features of the multivariate vector for which the fourth moment will be computed

Value

Returns the fourth moment.

Author(s)

Edgar Acuna

See Also

mo3, mardia

Examples

data(my.iris)
mo4(my.iris)

[Package dprep version 2.1 Index]