my.iris {dprep}	R Documentation

The Iris dataset

Description

This is perhaps the best known database to be found in the pattern recognition literature. Fisher's paper is a classic in the field and is referenced frequently to this day. (See Duda & Hart, for example.) The data set contains 3 classes of 50 instances each, where each class refers to a type of iris plant. The Setosa class is linearly separable from the other two classes. The last two classes are NOT linearly separable from each other.

Usage

data(my.iris)

Format

A dataframe with 150 observations on the following 5 variables.

V1

sepal length in cm

V2

sepal width in cm

V3

petal length in cm

V4

petal width in cm

V5

class: Iris Setosa(1), Iris Versicolor(2),Iris Virginica(3)

Source

The UCI Machine Learning Database Repository at:

ftp://ftp.ics.uci.edu/pub/machine-learning-databases
http://www.ics.uci.edu/~mlearn/MLRepository.html

References

Fisher, R. A. (1936) The use of multiple measurements in taxonomic problems. Annals of Eugenics, Vol 7, Part II, 179–188.

Examples

#----Testing multivariate normality----
data(my.iris)
mardia(my.iris)

[Package dprep version 2.1 Index]