near2 {dprep}	R Documentation

Auxiliary function for the reliefcat function

Description

This function finds the instance in the data matrix that is closest to a given instance x. The attributes can be either continuous or nominal.

Usage

near2(x, data, vnom)

Arguments

x 	 A given instance

data 	 The name of the dataset

vnom 	 A vector of indexes for nominal attributes

Author(s)

Edgar Acuna

See Also

relief,near1

[Package dprep version 2.1 Index]