outbox {dprep}	R Documentation

Detecting outliers through boxplots of the features.

Description

This function detects univariate outliers simultaneously using boxplots of the features.

Usage

outbox(data, nclass)

Arguments

data 	 The dataset to be explored for outlier detection.

nclass 	 A value representing the class that will be explored.

Details

The function also displays a plot containing a boxplot for of the variables.

Value

out1 	 A list of the indices of the observations that are outside the extremes of the boxplot. The indices are given in a table format representing the number of columns in which the observation was identified as an outlier.

Author(s)

Edgar Acuna

Examples

#---- Identifying outliers in diabetes-class1 with boxplots----
data(diabetes)
outbox(diabetes,nclass=1)

[Package dprep version 2.1 Index]