parallelplot {dprep}	R Documentation

Parallel Coordinate Plot

Description

Constructs a parallel coordinate plot for a data set with classes in last column.

Usage

parallelplot(x, name = "", comb = -1, class = 0, obs = rep(0, 0), col = 2, lty = 1, ...)

Arguments

x 	 A matrix of numerical values with classes in last column

name 	 The name of data set as will appear in the graph title

comb 	 An integer that represents the number of one of the possible combinations for the columns of this matrix.

class 	 A value representing the class number to which the plot should be limited

obs 	 A list of one or more row numbers that are to be highlighted in the plot

col 	 A value that provides a choice of color for the plot (if plotting only one class)

lty 	 A value that provides a choice of line width for the plot (if plotting only one class)

... 	 Additional arguments for the matplot function

Details

This plot is not recommended for a large number of features (say more than 50). If comb=0, all distinct combinations of columns are graphed. If comb=-1 (default), the attributes are plotted in their original order, else comb should be equal to an integer that represents the number of one of the possible combinations for the columns of this matrix.

Value

A parallel coordinate plot of the data is produced.

Author(s)

Caroline Rodriguez

References

Wegman, E. (1990), Hyperdimensional data analysis using parallel coordinates, Journal of the American Statistical Association,85,664-675

See Also

starcoord, surveyplot

Examples

#---Parallel Coordinate Plot----
data(bupa)
parallelplot(bupa,"Bupa Dataset")
parallelplot(bupa,"Bupa Dataset",comb=0)
#parallelplot(bupa,"Bupa Dataset",comb=1,c(1,22,50))

[Package dprep version 2.1 Index]