pp.golub {dprep}	R Documentation

The preprocessed Golub's dataset

Description

This is the preprocessed Golub's dataset, where the training and the learning set are first joined and then preprocessed according to the Dudoit et al.'s paper.

Usage

data(pp.golub)

Format

A data frame with 72 observations on 3572 variables.

Source

This data set is available at:

http://www-genome.wi.mit.edu/cgi-bin/cancer/datasets.cgi
http://www.bioconductor.org

References

Dudoit, S., Fridlyand, J. & Speed, T. P. (2002) Comparison of discrimination methods for the classification of tumors suing gene expression data. J Am Stat Assoc, 97 (457), 77–87.

Golub, T. R., Slonim, D. K., Tamayo, P., Huard, C., Gaasenbeek, M., Mesirov, J. P., Coller, H., Loh, M. L., Downing, J. R., Caligiuri, M. A., Bloomfield, C. D. & Lander, E. S. (1999) Molecular classification of cancer: class discovery and class prediction by gene expression monitoring. Science, 286, 531–537.

Examples

#----z-score Normalization---
data(pp.golub)
rangenorm(pp.golub,"znorm")

[Package dprep version 2.1 Index]