reachability {dprep}	R Documentation

Function for computing the reachability measure in the LOF algortihm

Description

This function computes the reachability measure for each instance of a dataset. This result is used later to compute the Local Outlyingness Factor.

Usage

reachability(distdata, k)

Arguments

distdata 	 The matrix of distances

k 	 The given number of neighbors

Author(s)

Caroline Rodriguez

See Also

maxlof

[Package dprep version 2.1 Index]