redundancy {dprep}	R Documentation

Finding the unique observations in a dataset along with their fequencies

Description

This function finds out the unique instances in a dataset along with their fequencies.

Usage

redundancy(data)

Arguments

data 	 The name of the dataset

Author(s)

Edgar Acuna

See Also

clean

[Package dprep version 2.1 Index]