reliefcat {dprep}	R Documentation

Feature selection by the Relief Algorithm for datasets with only nominal features

Description

This function applies the RELIEF Algorithm to datasets containing nominal attributes.

Usage

reliefcat(data, nosample, threshold, vnom)

Arguments

data 	 The name of the dataset

nosample 	 The size of the sample drawn and used to update the relevance of each feature

threshold 	 The threshold for choosing the relevant features

vnom 	 A vector of indices indicating the nominal features

Author(s)

Edgar Acuna

See Also

relief

[Package dprep version 2.1 Index]