robout {dprep}	R Documentation

Outlier Detection with Robust Mahalonobis distance

Description

This function finds the outliers of a dataset using robust versions of the Mahalanobis distance.

Usage

robout(data, nclass, meth = c("mve", "mcd"), rep = 10, 
plot = TRUE)

Arguments

data 	 The dataset for which outlier detection will be carried out.

nclass 	 An integer value that represents the class to detect for outliers

meth 	 The method used to compute the Mahalanobis distance, "mve"=minimum volume estimator, "mcd"=minimum covariance determinant

rep 	 Number of repetitions

plot 	 A boolean value to turn on and off the scatter plot of the Mahalanobis distances

Details

Requires uses cov.rob function from the MASS library.

Value

top1 	 Index of observations identified as top outliers by frequency of selection

topout 	 Index of observations identified as possible outliers by outlyingness measure

outme 	 Index of observations and their outlyingness measures

Author(s)

Edgar Acuna

References

Rousseeuw, P, and Leroy, A. (1987). Robust Regression and outlier detection. John Wiley & Sons. New York.

Atkinson, A. (1994). Fast very robust methods for the detection of multiple outliers. Journal of the American Statistical Association, 89:1329-1339.

Examples

#---- Outlier Detection in bupa-class 1 using MCD
data(bupa)
robout(bupa,1,"mcd")

[Package dprep version 2.1 Index]