row.matches {dprep}	R Documentation

Finding rows in a matrix equal to a given vector

Description

This function finds instances in a data matrix that are equal to a given instance.

Usage

row.matches(y, X)

Arguments

y 	 A given instance

X 	 A given data matrix

Details

This function was found in the CRAN mailing list. It seems to be authored by B. Venables

See Also

redundancy

[Package dprep version 2.1 Index]