sfs1 {dprep}	R Documentation

One-step sequential forward selection

Description

This function computes one-step of the sequential forward selection procedure.

Usage

sfs1(data, indic, correcto, kvec, method = c("lda", "knn",
 "rpart"))

Arguments

data 	 Name of the dataset to be used.

indic 	 A vector of 0-1 values.

correcto 	 The recognition rate in the previuos step.

kvec 	 The number of neighbors to be used by the knn classifier

method 	 The classifier to be used to select the best features.

Author(s)

Edgar Acuna

See Also

sffs

[Package dprep version 2.1 Index]