signorm {dprep}	R Documentation

Sigmoidal Normalization

Description

Function that performs sigmoidal normalization.

Usage

signorm(data)

Arguments

data 	 The dataset to be normalized, including classes

Details

This method transforms the input data nonlinearly into [-1,1] using a sigmoid function. The original data is first centered about the mean, and then mapped to the almost linear region of the sigmoid. Is especially appropriate when outlying values are present.

Removes classes before normalization, and returns the normalized data set complete with classes rejoined.

Value

sigdata 	 Original dataset normalized

Author(s)

Caroline Rodriguez and Edgar Acuna

References

Hann, J., Kamber, M. (2000). Data Mining: Concepts and Techniques. Morgan Kaufman Publishers.

Examples

#---- Sigmoidal Normalization ---
data(vehicle)
vehicle.signorm=signorm(vehicle)
op=par(mfrow=c(2,1))
plot(vehicle[,1])
plot(vehicle.signorm[,1])
par(op)

[Package dprep version 2.1 Index]