starcoord {dprep}	R Documentation

The star coordinates plot

Description

This function displays a star coordinates plot introduced by Kondogan (2001).

Usage

starcoord(data, main = NULL, class = FALSE, outliers=NULL, vars = 0, 
scale = 1, cex = 0.8, lwd = 0.25, lty = par("lty"))

Arguments

data 	 The dataset

main 	 The title of the plot

class 	 This logical variable is TRUE for supervised data and FALSE for unsupervised data

outliers 	 The instances to be highlighted as potential outliers

vars 	 The variables to be scaled

scale 	 The scale factor

cex 	 A numerical value giving the amount by which plotting text and symbols should be scaled.

lwd 	 The width of the lines representing the axis

lty 	 The type of the lines representing the axis

Details

This plot is not recommended for a large number of features (say more than 50).

Value

Returns a Star Coordinates Plot of the data matrix

Author(s)

Edgar Acuna and Shiyun Wen

References

E. Kandogan (2001). Visualizing multidimensional clusters, Trends, and Outliers, using star coordinates. Proceedings of KDD 2001.

See Also

parallelplot, surveyplot

Examples

data(bupa)
starcoord(bupa, main="Bupa Dataset", class=TRUE, outliers=NULL,vars=0, scale=1,
cex=0.8, lwd = 0.25, lty = par("lty"))

[Package dprep version 2.1 Index]