surveyplot {dprep}	R Documentation

Surveyplot

Description

This function creates and displays a surveyplot of a dataset for a classification matrix

Usage

surveyplot(datos, dataname = "", orderon = 0, class = 0, 
obs = rep(0, 0), lwd = 1)

Arguments

datos 	 A matrix of values for supervised classification

dataname 	 dataname Name of data set to appear in plot title

orderon 	 orderon Column number by which to order the dataset

class 	 class Class for which to limit plotting

obs 	 obs List of observations to be highlighted

lwd 	 lwd Value to control width of the line

Details

This plot is not recommended for a large number of features (say more than 50)

Value

Returns a surveyplot of the data matrix

Note

This plot is a mix between the survey plot presented in Fayyad and a permutation matrix.

Author(s)

Caroline Rodriguez

References

Fayyad, et al. (2001) Information Visualization in Data Mining and Knowledge Discovery

See Also

parallelplot, starcoord

Examples

#----Surveyplot examples
data(bupa)
surveyplot(bupa,"Bupa Dataset")
surveyplot(bupa,"Bupa Dataset",orderon=1,obs=c(6,74,121))

[Package dprep version 2.1 Index]