top {dprep}	R Documentation

Auxiliary function for Bay's Ouylier Detection Algorithm

Description

Function that finds the number of candidate outliers requested by the user.

Usage

top(O, neighbors, n)

Arguments

O 	 An n x 1 matrix with the score function from k nearest neighbors

neighbors 	 The number of neighbors to be considered

n 	 The number of top outliers to search for.

Author(s)

Caroline Rodriguez

See Also

baysout

[Package dprep version 2.1 Index]